+++ src/components/element/Typography.tsx
@@ -1,18 +1,14 @@
 import { ReactNode } from "react";
 
-import { dynamicTextAlign } from "@/style/dynamicPrams/typography/dynamicTextAlign";
-import { dynamicTextSize, TextSize } from "@/style/dynamicPrams/typography/dynamicTextSize";
+import { dynamicTextAlign, dynamicTextSize, TextAlign, TextSize } from "@/style/dynamicPrams/typography";
 
 interface Props {
     size?: TextSize;
-    textAlign?: "start" | "center" | "end";
+    textAlign?: TextAlign;
     className?: string;
     children: ReactNode;
 }
 
-/**
- * textColor, margin and padding are need to passed as tailwind's value.
- */
 export default function Typography({ size, textAlign, className, children }: Props) {
     const appliedStyle = `${dynamicTextSize[`${size ? size : "default"}`]} ${dynamicTextAlign[`${textAlign ? textAlign : "start"}`]} ${className}`;
     return <p className={appliedStyle}>{children}</p>;
