+++ src/style/dynamicPrams/grid/index.ts
@@ -1,0 +1,4 @@
+export * from "./dynamicColSpan";
+export * from "./dynamicGridCols";
+export * from "./dynamicGridRows";
+export * from "./dynamicRowSpan";
