+++ src/style/dynamicPrams/typography/index.ts
@@ -1,0 +1,2 @@
+export * from "./dynamicTextAlign";
+export * from "./dynamicTextSize";
